+++ src/math_extra.zig
@@ -46,3 +46,25 @@ pub fn saturatingSub(comptime T: type, a: T, b: T) T {
         else => @compileError("Only integer and float types supported"),
     }
 }
+
+pub fn saturatingAdd(comptime T: type, a: T, b: T) T {
+    switch (@typeInfo(T)) {
+        .int,
+        .comptime_int,
+        => return math.add(T, a, b) catch {
+            return if (a < b)
+                0
+            else
+                math.maxInt(T);
+        },
+        .float,
+        .comptime_float,
+        => return math.add(T, a, b) catch {
+            return if (a < b)
+                0
+            else
+                math.floatMax(T);
+        },
+        else => @compileError("Only integer and float types supported"),
+    }
+}