+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow infinite range in `fromRangeStep`
 - Panic instead of asserting where applicable
 
+### Fixed
+
+- Overflow error handling in `Chain`
+
 ## [0.5.0] - 2025-06-02
 
 ### Added